requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
click>=8.1.0

# AI/ML dependencies (add specific ones as needed)
openai>=1.0.0
anthropic>=0.7.0
groq>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiohttp>=3.8.0
asyncio